+++ app/shared/modals/reservation-cancel/ReservationCancelModalContainer.js
@@ -7,6 +7,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import Toggle from 'react-toggle';
 
+import { getHasOnlinePaymentSupport } from '../../../../src/domain/resource/utils';
 import { deleteReservation } from '../../../actions/reservationActions';
 import { closeReservationCancelModal } from '../../../actions/uiActions';
 import CompactReservationList from '../../compact-reservation-list/CompactReservationList';
@@ -20,16 +21,6 @@ class UnconnectedReservationCancelModalContainer extends Component {
     this.handleCancel = this.handleCancel.bind(this);
     this.handleCheckbox = this.handleCheckbox.bind(this);
     this.state = { checkboxDisabled: null };
-  }
-
-  componentDidMount() {
-    const {
-      resource,
-    } = this.props;
-
-    hasProducts(resource)
-      ? this.setState({ checkboxDisabled: false })
-      : this.setState({ checkboxDisabled: true });
   }
 
   handleCancel() {
@@ -65,6 +56,8 @@ class UnconnectedReservationCancelModalContainer extends Component {
       show,
       t,
     } = this.props;
+
+    const resourceHasOnlinePaymentSupport = getHasOnlinePaymentSupport(resource);
 
     return (
       <Modal
@@ -106,8 +99,18 @@ class UnconnectedReservationCancelModalContainer extends Component {
           {!cancelAllowed
             && (
             <div>
-              <p>{t('ReservationCancelModal.cancelNotAllowedInfo')}</p>
-              <p><FormattedHTMLMessage id="ReservationCancelModal.takeIntoAccount" /></p>
+              {resourceHasOnlinePaymentSupport && (
+                <>
+                  <p>{t('ReservationCancelModal.cancelNotAllowedInfoOnlinePayment.1')}</p>
+                  <p>{t('ReservationCancelModal.cancelNotAllowedInfoOnlinePayment.2')}</p>
+                </>
+              )}
+              {!resourceHasOnlinePaymentSupport && (
+                <>
+                  <p>{t('ReservationCancelModal.cancelNotAllowedInfo')}</p>
+                  <p><FormattedHTMLMessage id="ReservationCancelModal.takeIntoAccount" /></p>
+                </>
+              )}
               <p className="responsible-contact-info">{resource.responsibleContactInfo}</p>
             </div>
             )