+++ src/domain/reservation/modal/__tests__/ReservationInfomationModal.test.js
@@ -51,4 +51,37 @@ describe('ReservationInformationModal', () => {
       expect(findButtonByLabel(wrapper, 'denyButton').length).toEqual(1);
     });
   });
+
+  describe('refund policy', () => {
+    test('rendered when resource is not null, when user is admin, when reservation is not a staff event and when the reservation has a price', () => {
+      const getRefundPolicy = (wrapper) => {
+        return wrapper.find({ id: 'refund-policy' });
+      };
+      const notAdmin = getWrapper({ isEditing: true, isAdmin: false });
+      const admin = getWrapper({
+        isEditing: true,
+        isAdmin: true,
+        reservation: {
+          ...mockReservation,
+          isStaffEvent: false,
+          begin: new Date(2017, 10, 1, 9, 0, 0, 0).toJSON(),
+          end: new Date(2017, 10, 1, 11, 0, 0, 0).toJSON(),
+        },
+        resource: {
+          products: [
+            {
+              price: {
+                type: 'per_period',
+                period: '01:00',
+                amount: 100,
+              },
+            },
+          ],
+        },
+      });
+
+      expect(getRefundPolicy(notAdmin).length).toEqual(0);
+      expect(getRefundPolicy(admin).length).toEqual(1);
+    });
+  });
 });